+++ index.ts
@@ -1,0 +1,109 @@
+
+export {
+  // convertValueToBoolean,
+  // displaySecondsAsMinutes,
+  addComputerLog,
+  addErrorLog,
+  addLog,
+  compareItemsForSorting,
+  compareObjectProperties,
+  constructCSVFile,
+  convertEnableDisableTrueFalse,
+  convertMilitaryTimeToStandardTime,
+  convertMinutesToSeconds,
+  convertNormalAbnormalTrueFalse,
+  convertNullEmptyString,
+  convertSpecialCharacters,
+  convertStandardTimeToMilitaryTime,
+  convertTemperature,
+  convertTimeToMinutes,
+  convertYesNoTrueFalse,
+  createDateFromString,
+  daysSince,
+  displayCommaBetweenTwoItems,
+  displayDate,
+  displayDateAndTime,
+  displayObjectData,
+  displayObjectDataTable,
+  displayObjectDataXML,
+  displaySpaceAfterComma,
+  displayTime,
+  displayValue,
+  displayYear,
+  emailFormat,
+  exportCSVFile,
+  filterBySelectedCheckbox,
+  filterDropdownOptions,
+  formatDate,
+  formatFloat,
+  formatInt,
+  formatLowerCase,
+  formatPhoneNumber,
+  formatSearchInput,
+  formatTitle,
+  formatToString,
+  formatToUSD,
+  formatTrim,
+  formatUpperCase,
+  generateHoursInterval,
+  generateRandomNumber,
+  generateRandomNumberDigits,
+  getBrowserData,
+  getCurrentDay,
+  getCurrentMonth,
+  getCurrentYear,
+  getDateTime,
+  getFilteredNames,
+  getFirstItem,
+  getLastItem,
+  getNumberOfDaysBetweenDates,
+  getObjectArrayUniqueProperty,
+  getParseInt,
+  getQueryStringData,
+  getYears,
+  groupObjectArrayByProperties,
+  hasDecimalPlaces,
+  hasEqualsProperty,
+  hasFalseProperty,
+  hasNonEmptyProperty,
+  hasTrueProperty,
+  isEmpty,
+  isEmptyArray,
+  isFutureDate,
+  isNonEmptyArray,
+  isValidURL,
+  isWholeNumber,
+  noFunctionAvailable,
+  parse,
+  parseDescription,
+  randomizeItems,
+  removeArticlesFromBeginning,
+  removeForwardSlashes,
+  removeHTML,
+  removeNonAlphanumericCharacters,
+  replaceSmartCharacters,
+  returnActiveClass,
+  sortObjectArrayByProperty,
+  sortObjectArrayByTwoProperties,
+  truncateText,
+  tryParseJSON,
+  validateMilitaryTime
+} from "./utilities/sharedFunctions.js";
+
+export {
+  allowLogging,
+  getFetchAuthorization,
+  inElectron,
+  inLearningObjectEcosystem,
+  isLocalDevelopment,
+  resolveBaseURL,
+  resolveRedirectURL,
+  // selectEnvironmentVariable,
+  showAuthentication,
+  showDemonstration,
+  showDevelopment,
+  showErrorMessage,
+  showLocalDevelopment,
+  showPlayground,
+  showStaging
+} from "./utilities/developerFunctions.js";
