+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "shared-functions",
-  "version": "1.2.1",
+  "type": "module",
+  "version": "1.2.2",
   "description": "Contains common functions that could apply to all applications developed.",
   "author": "Orbis Education",
   "copyrightYear": "2026",
@@ -9,11 +10,43 @@
     "url": "https://github.com/orbis-education/shared-functions"
   },
   "private": true,
-  "main": "index.js",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "clean": "rimraf dist",
+    "build": "npm run clean && tsc",
+    "prepare": "npm run build",
+    "lint": "eslint --max-warnings=0 utilities index.ts",
+    "lint:fix": "eslint --max-warnings=0 lib utilities index.ts --fix",
+    "format": "prettier --write \"**/*.{ts,json}\"",
+    "format:check": "prettier --check \"**/*.{ts,json}\"",
+    "typecheck": "tsc --noEmit"
   },
   "dependencies": {
     "html-react-parser": "^5.2.17"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.39.2",
+    "@types/node": "^25.1.0",
+    "@typescript-eslint/eslint-plugin": "^8.54.0",
+    "@typescript-eslint/parser": "^8.54.0",
+    "eslint": "^9.39.2",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "globals": "^17.3.0",
+    "prettier": "^3.4.2",
+    "rimraf": "^6.1.3",
+    "tsdown": "^0.21.10",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.54.0"
   }
 }
