+++ utilities/sharedFunctions.ts
@@ -1,0 +1,2376 @@
+import parseHTML, { type HTMLReactParserOptions } from "html-react-parser";
+
+export type PrimitiveValue =
+  | string
+  | number
+  | boolean
+  | null
+  | undefined;
+
+export type UnknownRecord = Record<string, unknown>;
+
+export type SortDirection = "asc" | "desc" | string;
+
+export const emailFormat = /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/;
+
+export const noFunctionAvailable = (): void => {
+
+  console.log("A function wasn't passed as a props when it needed to be.");
+
+};
+
+export const isEmpty = (value: unknown): boolean => {
+
+  // * https://stackoverflow.com/questions/4597900/checking-something-isempty-in-javascript -- 03/06/2021 MF
+  // * https://stackoverflow.com/questions/5515310/is-there-a-standard-function-to-check-for-null-undefined-or-blank-variables-in -- 03/06/2021 MF
+
+  return value === undefined || value === null || (typeof value === "object" && Object.keys(value).length === 0) || (typeof value === "string" && value.trim().length === 0);
+
+  // * Returns true -- 03/06/2021 MF
+  // console.log("isEmpty(\"\")", isEmpty(""));
+  // console.log("isEmpty(null)", isEmpty(null));
+  // console.log("isEmpty(undefined)", isEmpty(undefined));
+  // console.log("isEmpty([])", isEmpty([]));
+  // console.log("isEmpty({})", isEmpty({}));
+
+  // * Returns false -- 03/06/2021 MF
+  // console.log("isEmpty(\"test\")", isEmpty("test"));
+  // console.log("isEmpty(5)", isEmpty(5));
+  // console.log("isEmpty(true)", isEmpty(true));
+  // console.log("isEmpty([\"test\"])", isEmpty(["test"]));
+  // console.log("isEmpty({test: \"test\"})", isEmpty({ test: "test" }));
+
+};
+
+
+export const getDateTime = () => {
+
+  // * Time returned does not consider the time zone without adjustments. -- 08/09/2021 MF
+  // * https://usefulangle.com/post/30/javascript-get-date-time-with-offset-hours-minutes -- 08/09/2021 MF
+
+  // * https://stackoverflow.com/questions/12413243/javascript-date-format-like-iso-but-local -- 08/09/2021 MF
+  let timezoneOffset = new Date().getTimezoneOffset() * 60 * 1000;
+
+  // return new Date().toLocaleString();
+  // return new Date().toLocaleString().slice(0, 19).replace("T", " ");
+  // return new Date().toISOString().slice(0, 19).replace("T", " ");
+  return new Date(new Date() - timezoneOffset).toISOString().slice(0, 19).replace("T", " ");
+
+};
+
+
+export const isNonEmptyArray = (arrayItem) => {
+
+  let nonEmptyArray = false;
+
+  if (Array.isArray(arrayItem) && arrayItem.length > 0) {
+
+    nonEmptyArray = true;
+
+  };
+
+  return nonEmptyArray;
+
+};
+
+
+// * Doesn't consistently work as expected. -- 09/09/2025 MF
+// export const isEmptyArray = (arrayItem) =>
+//   Array.isArray(arrayItem) && arrayItem.length === 0;
+
+export const isEmptyArray = (arrayItem) => {
+
+  let emptyArray = true;
+
+  if (Array.isArray(arrayItem) && arrayItem.length > 0) {
+
+    emptyArray = false;
+
+  };
+
+  return emptyArray;
+
+};
+
+
+export const getFirstItem = (arrayItem) => {
+
+  let firstItem = {};
+
+  if (isNonEmptyArray(arrayItem) && !isEmpty(arrayItem[0])) {
+
+    firstItem = arrayItem[0];
+
+  };
+
+  return firstItem;
+
+};
+
+
+export const getLastItem = (arrayItem) => {
+
+  let lastItem = {};
+
+  if (isNonEmptyArray(arrayItem) && !isEmpty(arrayItem[arrayItem.length - 1])) {
+
+    lastItem = arrayItem[arrayItem.length - 1];
+
+  };
+
+  return lastItem;
+
+};
+
+
+export const displayValue = (variableValue) => {
+
+  let displayValue = "";
+
+  if (!isEmpty(variableValue)) {
+
+    if (variableValue === true) {
+
+      displayValue = "True";
+
+    } else if (variableValue === false) {
+
+      displayValue = "False";
+
+    } else if (variableValue instanceof Date) {
+
+      displayValue = variableValue.toLocaleString();
+
+    } else {
+
+      displayValue = variableValue;
+
+    };
+
+  } else {
+
+    displayValue = "Value is undefined or null.";
+
+  };
+
+  return displayValue;
+
+};
+
+
+export const displaySpaceAfterComma = (text) => {
+
+  let displayText = text;
+
+  if (!isEmpty(text)) {
+
+    displayText = text.replaceAll(",", ", ");
+
+  };
+
+  return displayText;
+
+};
+
+
+export const removeForwardSlashes = (text) => {
+
+  let displayText = text;
+
+  if (!isEmpty(text)) {
+
+    displayText = formatToString(text).replace(/\//g, "");
+
+  };
+
+  return displayText;
+
+};
+
+
+export const tryParseJSON = (jsonString) => {
+
+  // * https://stackoverflow.com/questions/3710204/how-to-check-if-a-string-is-a-valid-json-string-in-javascript-without-using-try -- 03/06/2021 MF
+  try {
+
+    let jsonData = JSON.parse(jsonString);
+
+    // * Handle non-exception-throwing cases: -- 03/05/2021
+    // * Neither JSON.parse(false) or JSON.parse(1234) throw errors, hence the type-checking, -- 03/05/2021
+    // * but... JSON.parse(null) returns null, and typeof null === "object", -- 03/05/2021
+    // * so we must check for that, too. Thankfully, null is falsey, so this suffices: -- 03/05/2021
+    if (jsonData && typeof jsonData === "object") {
+
+      return jsonData;
+
+    };
+
+  }
+  catch (error) {
+
+    // * Don't display this error in the console. This function is already returning false if the JSON file is not in the correct format. -- 03/06/2021 MF
+    // console.log("tryParseJSON error", error);
+
+  };
+
+  return false;
+
+};
+
+
+export const displayObjectData = (ObjectData) => {
+
+  let objectDataString = JSON.stringify(ObjectData);
+
+  if (!isEmpty(objectDataString)) {
+
+    objectDataString = objectDataString.replaceAll("\\", "");
+
+    objectDataString = objectDataString.replaceAll("[{\"", "<p><strong>");
+    objectDataString = objectDataString.replaceAll("\"},{\"", "</p><p><strong>");
+    objectDataString = objectDataString.replaceAll("\"}]", "</p>");
+
+    objectDataString = objectDataString.replaceAll("{\"", "<p><strong>");
+    objectDataString = objectDataString.replaceAll("\"}", "</p>");
+
+    objectDataString = objectDataString.replaceAll("\":\"", "</strong> = ");
+    objectDataString = objectDataString.replaceAll("\":", "</strong> = ");
+
+    objectDataString = objectDataString.replaceAll("\",\"", "</p><p><strong>");
+    objectDataString = objectDataString.replaceAll(",\"", "</p><p><strong>");
+
+    objectDataString = objectDataString.replaceAll("},", "");
+
+    objectDataString = objectDataString.replaceAll("[]", "");
+
+    objectDataString = objectDataString.replaceAll("[\"", "");
+    objectDataString = objectDataString.replaceAll("\"]", "");
+
+    objectDataString = objectDataString.replaceAll("[", "");
+    objectDataString = objectDataString.replaceAll("]", "");
+    objectDataString = objectDataString.replaceAll("{", "");
+    objectDataString = objectDataString.replaceAll("}", "");
+
+    objectDataString = objectDataString.replace(/<strong>(.*?)<\/strong>/g, (match) => { return formatTitle(match); });
+
+  };
+
+  return (objectDataString);
+
+};
+
+
+// TODO: Account for JSON data embedded in JSON data. -- 09/18/2024 MF
+export const displayObjectDataTable = (ObjectData) => {
+
+  let objectDataString = JSON.stringify(ObjectData);
+
+  if (!isEmpty(objectDataString)) {
+
+    objectDataString = objectDataString.replaceAll("\\", "");
+
+    objectDataString = objectDataString.replaceAll("[{\"", "<tr><th>");
+    objectDataString = objectDataString.replaceAll("\"},{\"", "</td><tr><th>");
+    objectDataString = objectDataString.replaceAll("\"}]", "</td></tr>");
+    // objectDataString = objectDataString.replaceAll("[{\\\"", "<tr><th>");
+    // objectDataString = objectDataString.replaceAll("\\\"},{\\\"", "</td><tr><th>");
+    // objectDataString = objectDataString.replaceAll("\\\"}]", "</td></tr>");
+
+    objectDataString = objectDataString.replaceAll("{\"", "<tr><th>");
+    objectDataString = objectDataString.replaceAll("\"}", "</td></tr>");
+    // objectDataString = objectDataString.replaceAll("{\\\"", "<tr><th>");
+    // objectDataString = objectDataString.replaceAll("\\\"}", "</td></tr>");
+
+    objectDataString = objectDataString.replaceAll("\":\"", "</th><td>");
+    objectDataString = objectDataString.replaceAll("\":", "</th><td>");
+    // objectDataString = objectDataString.replaceAll("\\\":\\\"", "</th><td>");
+    // objectDataString = objectDataString.replaceAll("\\\":", "</th><td>");
+
+    objectDataString = objectDataString.replaceAll("\",\"", "</td><tr><th>");
+    objectDataString = objectDataString.replaceAll(",\"", "</td><tr><th>");
+    // objectDataString = objectDataString.replaceAll("\\\",\\\"", "</td><tr><th>");
+    // objectDataString = objectDataString.replaceAll(",\\\"", "</td><tr><th>");
+
+    objectDataString = objectDataString.replaceAll("},", "");
+
+    objectDataString = objectDataString.replaceAll("[]", "");
+
+    objectDataString = objectDataString.replaceAll("[\"", "");
+    objectDataString = objectDataString.replaceAll("\"]", "");
+    // objectDataString = objectDataString.replaceAll("[\\\"", "");
+    // objectDataString = objectDataString.replaceAll("\\\"]", "");
+
+    objectDataString = objectDataString.replaceAll("[", "");
+    objectDataString = objectDataString.replaceAll("]", "");
+    objectDataString = objectDataString.replaceAll("{", "");
+    objectDataString = objectDataString.replaceAll("}", "");
+
+
+    objectDataString = objectDataString.replace(/<th>(.*?)<\/th>/g, (match) => { return formatTitle(match); });
+
+    objectDataString = `<div class="table-container"><table><thead><tr class="sr-only"><th>Property Name</th><th>Value</th></tr></thead><tbody>${objectDataString}</tbody></table></div>`;
+
+  };
+
+  return (objectDataString);
+
+};
+
+
+export const displayObjectDataXML = (ObjectData) => {
+
+  let objectDataString = JSON.stringify(ObjectData);
+
+  if (!isEmpty(objectDataString)) {
+
+    objectDataString = objectDataString.replaceAll("\\", "");
+
+    objectDataString = objectDataString.replaceAll("[{\"", "<category>");
+    objectDataString = objectDataString.replaceAll("\"},{\"", "</data><category>");
+    objectDataString = objectDataString.replaceAll("\"}]", "<data>");
+
+    objectDataString = objectDataString.replaceAll("{\"", "<category>");
+    objectDataString = objectDataString.replaceAll("\"}", "<data>");
+
+    objectDataString = objectDataString.replaceAll("\":\"", "</category><data>");
+    objectDataString = objectDataString.replaceAll("\":", "</category><data>");
+
+    objectDataString = objectDataString.replaceAll("\",\"", "</data><category>");
+    objectDataString = objectDataString.replaceAll(",\"", "</data><category>");
+
+    objectDataString = objectDataString.replaceAll("},", "");
+
+    objectDataString = objectDataString.replaceAll("[]", "");
+
+    objectDataString = objectDataString.replaceAll("[\"", "");
+    objectDataString = objectDataString.replaceAll("\"]", "");
+
+    objectDataString = objectDataString.replaceAll("[", "");
+    objectDataString = objectDataString.replaceAll("]", "");
+    objectDataString = objectDataString.replaceAll("{", "");
+    objectDataString = objectDataString.replaceAll("}", "");
+
+  };
+
+  return (objectDataString);
+
+};
+
+
+export const getCurrentDay = () => {
+
+  // * Time returned does not consider the time zone without adjustments. -- 08/09/2021 MF
+  // * https://usefulangle.com/post/30/javascript-get-date-time-with-offset-hours-minutes -- 08/09/2021 MF
+
+  // * https://stackoverflow.com/questions/12413243/javascript-date-format-like-iso-but-local -- 08/09/2021 MF
+  let timezoneOffset = new Date().getTimezoneOffset() * 60 * 1000;
+
+  // return new Date().toLocaleString();
+  // return new Date().toLocaleString().slice(0, 19).replace("T", " ");
+  // return new Date().toISOString().slice(0, 19).replace("T", " ");
+  // return new Date().getDate();
+  return new Date(new Date() - timezoneOffset).getDate();
+
+};
+
+
+export const getCurrentMonth = () => {
+
+  // * Time returned does not consider the time zone without adjustments. -- 08/09/2021 MF
+  // * https://usefulangle.com/post/30/javascript-get-date-time-with-offset-hours-minutes -- 08/09/2021 MF
+
+  // * https://stackoverflow.com/questions/12413243/javascript-date-format-like-iso-but-local -- 08/09/2021 MF
+  let timezoneOffset = new Date().getTimezoneOffset() * 60 * 1000;
+
+  // return new Date().toLocaleString();
+  // return new Date().toLocaleString().slice(0, 19).replace("T", " ");
+  // return new Date().toISOString().slice(0, 19).replace("T", " ");
+  // return new Date().getMonth() + 1;
+  return new Date(new Date() - timezoneOffset).getMonth() + 1;
+
+};
+
+
+export const getCurrentYear = () => {
+
+  // * Time returned does not consider the time zone without adjustments. -- 08/09/2021 MF
+  // * https://usefulangle.com/post/30/javascript-get-date-time-with-offset-hours-minutes -- 08/09/2021 MF
+
+  // * https://stackoverflow.com/questions/12413243/javascript-date-format-like-iso-but-local -- 08/09/2021 MF
+  let timezoneOffset = new Date().getTimezoneOffset() * 60 * 1000;
+
+  // return new Date().toLocaleString();
+  // return new Date().toLocaleString().slice(0, 19).replace("T", " ");
+  // return new Date().toISOString().slice(0, 19).replace("T", " ");
+  // return new Date().getFullYear();
+  return new Date(new Date() - timezoneOffset).getFullYear();
+
+};
+
+
+export const displayDate = (dateToDisplay, removeLeadingZeroes) => {
+
+  let newDisplayDate = "";
+
+  if (!isEmpty(dateToDisplay)) {
+
+    // * Year
+    let yyyy = formatToString(dateToDisplay).substring(0, 4);
+    // * Month
+    let mm = formatToString(dateToDisplay).substring(5, 7);
+    // * Day
+    let dd = formatToString(dateToDisplay).substring(8, 10);
+
+    newDisplayDate = mm + "/" + dd + "/" + yyyy;
+
+    if (!isEmpty(newDisplayDate) && removeLeadingZeroes === true) {
+
+      newDisplayDate = newDisplayDate.replace(/\b0/g, "");
+
+    };
+
+  };
+
+  return newDisplayDate;
+
+};
+
+
+export const displayDateAndTime = (dateToDisplay, removeLeadingZeroes) => {
+
+  let newDisplayDateAndTime = "";
+
+  if (!isEmpty(dateToDisplay)) {
+
+    // * Year
+    let yyyy = formatToString(dateToDisplay).substring(0, 4);
+    // * Month
+    let mm = formatToString(dateToDisplay).substring(5, 7);
+    // * Day
+    let dd = formatToString(dateToDisplay).substring(8, 10);
+
+    // // * Hour
+    // let hour = formatToString(dateToDisplay).substring(11, 12);
+    // // * Minute
+    // let minute = formatToString(dateToDisplay).substring(15, 16);
+
+    // * Time
+    let time = formatToString(dateToDisplay).substring(11, 16);
+
+    newDisplayDateAndTime = mm + "/" + dd + "/" + yyyy + " " + time;
+
+    if (!isEmpty(newDisplayDateAndTime) && removeLeadingZeroes === true) {
+
+      newDisplayDateAndTime = newDisplayDateAndTime.replace(/\b0/g, "");
+
+    };
+
+  };
+
+  return newDisplayDateAndTime;
+
+};
+
+
+export const displayYear = (dateToDisplay) => {
+
+  let newDisplayDate = "";
+
+  if (!isEmpty(dateToDisplay)) {
+
+    // * Year
+    let yyyy = formatToString(dateToDisplay).substring(0, 4);
+    // * Month
+    // let mm = formatToString(dateToDisplay).substring(5, 7);
+    // * Day
+    // let dd = formatToString(dateToDisplay).substring(8, 10);
+
+    // newDisplayDate = mm + "/" + dd + "/" + yyyy;
+
+    newDisplayDate = yyyy;
+
+  };
+
+  return newDisplayDate;
+
+};
+
+
+export const daysSince = (dateToCompare) => {
+
+  // * https://stackoverflow.com/questions/12986068/how-to-calculate-number-of-days-between-today-and-given-date-and-code-for-gettim -- 10/18/2021 MF
+
+  let newDaysSince = 0;
+
+  if (!isEmpty(dateToCompare)) {
+
+    let today = new Date();
+    let compareDate = new Date(dateToCompare);
+    let timeInMilliseconds = compareDate.getTime() - today.getTime();
+
+    newDaysSince = Math.abs(Math.ceil(timeInMilliseconds / (1000 * 60 * 60 * 24)));
+
+  };
+
+  return newDaysSince;
+
+};
+
+
+export const hasNonEmptyProperty = (objectItem, propertyName) => {
+
+  // * This function is intended for use in ternary statements so that there aren't lines and lines of if statement structures or to see if the object has a property available with a value. -- 06/09/2022 MF
+
+  let nonEmptyProperty = false;
+
+  if (typeof objectItem === "object") {
+
+    if (Object.hasOwn(objectItem, propertyName) && !isEmpty(objectItem[propertyName])) {
+
+      nonEmptyProperty = true;
+
+    };
+
+  };
+
+  return nonEmptyProperty;
+
+};
+
+
+export const hasEqualsProperty = (objectItem, propertyName, value) => {
+
+  // * This function is intended for use in ternary statements so that there aren't lines and lines of if statement structures. -- 06/09/2022 MF
+
+  let equalsProperty = false;
+
+  if (typeof objectItem === "object") {
+
+    if (Object.hasOwn(objectItem, propertyName) && !isEmpty(objectItem[propertyName]) && objectItem[propertyName] === value) {
+
+      equalsProperty = true;
+
+    };
+
+  };
+
+  return equalsProperty;
+
+};
+
+
+export const hasTrueProperty = (objectItem, propertyName) => {
+
+  // * This function is intended for use in ternary statements so that there aren't lines and lines of if statement structures. -- 06/09/2022 MF
+
+  let trueProperty = false;
+
+  if (typeof objectItem === "object") {
+
+    if (Object.hasOwn(objectItem, propertyName) && objectItem[propertyName] === true) {
+
+      trueProperty = true;
+
+    };
+
+  };
+
+  return trueProperty;
+
+};
+
+
+export const hasFalseProperty = (objectItem, propertyName) => {
+
+  // * This function is intended for use in ternary statements so that there aren't lines and lines of if statement structures. -- 06/09/2022 MF
+
+  let falseProperty = false;
+
+  if (typeof objectItem === "object") {
+
+    if (Object.hasOwn(objectItem, propertyName) && objectItem[propertyName] === false) {
+
+      falseProperty = true;
+
+    };
+
+  };
+
+  return falseProperty;
+
+};
+
+
+export const convertSpecialCharacters = (value) => {
+
+  // * https://stackoverflow.com/questions/1787322/what-is-the-htmlspecialchars-equivalent-in-javascript/4835406#4835406 -- 12/28/2021 MF
+
+  let newValue = value;
+
+  if (!isEmpty(value)) {
+
+    newValue = newValue.replace(/&/g, "&amp;");
+    newValue = newValue.replace(/</g, "&lt;");
+    newValue = newValue.replace(/>/g, "&gt;");
+    newValue = newValue.replace(/"/g, "&quot;");
+    newValue = newValue.replace(/'/g, "&#039;");
+
+  };
+
+  return newValue;
+
+};
+
+
+export const truncateText = (text, limit) => {
+
+  // * https://stackoverflow.com/questions/4700226/i-want-to-truncate-a-text-or-line-with-ellipsis-using-javascript -- 03/06/2021 MF
+
+  if (!isEmpty(text) && text.length > limit) {
+
+    for (let i = limit; i > 0; i--) {
+
+      if (text.charAt(i) === " " && (text.charAt(i - 1) !== "," || text.charAt(i - 1) !== "." || text.charAt(i - 1) !== ";")) {
+
+        return text.substring(0, i) + "...";
+
+      };
+
+    };
+
+    return text.substring(0, limit) + "...";
+
+  } else {
+
+    return text;
+
+  };
+
+};
+
+
+export const validateMilitaryTime = (timeEntered) => {
+
+  // * Time in 24 clock, no colon -- 03/05/2021 MF
+
+  // * all digits-- 03/05/2021 MF
+  // * length is 4
+  // * first digit is either a 0 or 1 or 2
+  // * second digit is either a 0 or 1 or 2 or 3 or 4 or 5 
+  // * third digit is either a 0 or 1 or 2 or 3 or 4 or 5 
+
+  // * Make sure that it is a string-- 03/05/2021 MF
+
+  let validTimeFormat = true;
+
+  let timeEnteredString = "";
+
+  if (typeof formatToString(timeEntered) === "string") {
+
+    timeEnteredString = formatToString(timeEntered);
+
+    timeEnteredString = timeEnteredString.trim();
+
+    if (timeEnteredString.length !== 4) {
+
+      validTimeFormat = false;
+
+    } else {
+
+      if (isNaN(parseInt(timeEnteredString.charAt(0))) || isNaN(parseInt(timeEnteredString.charAt(1))) || isNaN(parseInt(timeEnteredString.charAt(2))) || isNaN(parseInt(timeEnteredString.charAt(3)))) {
+
+        validTimeFormat = false;
+
+      };
+
+      if (timeEnteredString.charAt(0) !== "0" && timeEnteredString.charAt(0) !== "1" && timeEnteredString.charAt(0) !== "2") {
+
+        validTimeFormat = false;
+
+      };
+
+      if (timeEnteredString.charAt(0) === "2" && timeEnteredString.charAt(1) !== "0" && timeEnteredString.charAt(1) !== "1" && timeEnteredString.charAt(1) !== "2" && timeEnteredString.charAt(1) !== "3" && timeEnteredString.charAt(1) !== "4") {
+
+        validTimeFormat = false;
+
+      };
+
+      if (timeEnteredString.charAt(0) === "2" && timeEnteredString.charAt(1) !== "4" && timeEnteredString.charAt(2) !== "0" && timeEnteredString.charAt(3) !== "0") {
+
+        validTimeFormat = false;
+
+      };
+
+      if (timeEnteredString.charAt(2) !== "0" && timeEnteredString.charAt(2) !== "1" && timeEnteredString.charAt(2) !== "2" && timeEnteredString.charAt(2) !== "3" && timeEnteredString.charAt(2) !== "4" && timeEnteredString.charAt(2) !== "5") {
+
+        validTimeFormat = false;
+
+      };
+
+    };
+
+  } else {
+
+    validTimeFormat = false;
+
+  };
+
+  return validTimeFormat;
+
+};
+
+
+export const convertTemperature = (temperatureScale, temperature) => {
+
+  // let temperatureFloat = parseFloat(formatTrim(temperature));
+  let temperatureFloat = parseFloat(temperature);
+  let temperatureConverted = "";
+
+  if (!isEmpty(temperatureFloat) && !isNaN(temperatureFloat)) {
+
+    if (formatLowerCase(temperatureScale) === "celsius") {
+
+      // * Based on (32°F - 32) x 5/9 = 0°C -- 07/29/2021 MF
+      temperatureConverted = ((temperatureFloat - 32) * 5 / 9).toFixed(2);
+
+    } else if (formatLowerCase(temperatureScale) === "fahrenheit") {
+
+      // * Based on (32°F - 32) x 5/9 = 0°C -- 07/29/2021 MF
+      temperatureConverted = (temperatureFloat * 9 / 5 + 32).toFixed(2);
+
+    };
+
+  } else {
+
+    temperatureConverted = "";
+
+  };
+
+  return temperatureConverted;
+
+};
+
+
+export const convertYesNoTrueFalse = (value) => {
+
+  if (isNaN(value) && value === "Yes") {
+
+    return true;
+
+  } else if (isNaN(value) && value === "No") {
+
+    return false;
+
+  } else if (value === true) {
+
+    return "Yes";
+
+  } else if (value === false) {
+
+    return "No";
+
+  } else {
+
+    return value;
+
+  };
+
+};
+
+
+export const convertNormalAbnormalTrueFalse = (value) => {
+
+  if (isNaN(value) && value === "Normal") {
+
+    return true;
+
+  } else if (isNaN(value) && value === "Abnormal") {
+
+    return false;
+
+  } else if (value === true) {
+
+    return "Normal";
+
+  } else if (value === false) {
+
+    return "Abnormal";
+
+  } else {
+
+    return value;
+
+  };
+
+};
+
+
+export const convertEnableDisableTrueFalse = (value) => {
+
+  if (isNaN(value) && value === "Enable") {
+
+    return true;
+
+  } else if (isNaN(value) && value === "Disable") {
+
+    return false;
+
+  } else if (value === true) {
+
+    return "Enable";
+
+  } else if (value === false) {
+
+    return "Disable";
+
+  } else {
+
+    return value;
+
+  };
+
+};
+
+
+export const convertNullEmptyString = (value) => {
+
+  // * TODO: Change this function so that it can handle if there are already empty string values in the database. -- 03/19/2021 MF
+  // * This can't be done in one function like this to handle both conversions because what if the database value is set to an empty string. -- 07/09/2021 MF
+
+  if (value === null) {
+
+    return "";
+
+  } else if (value === undefined) {
+
+    return "";
+
+  } else if (value === "NaN") {
+
+    return null;
+
+  } else if (isNaN(value) && typeof value === "number") {
+
+    return null;
+
+  } else if (isNaN(value) && value === "") {
+
+    return null;
+
+  } else if (value === "") {
+
+    return null;
+
+  } else {
+
+    return value;
+
+  };
+
+};
+
+
+export const isWholeNumber = (value) => {
+
+  // * Make sure that the value isn't empty. isNaN counts empty values as a number. -- 09/11/2023 MF
+  if (isEmpty(value) || isNaN(formatTrim(value))) {
+
+    return false;
+
+  } else {
+
+    // * https://www.w3resource.com/javascript-exercises/javascript-math-exercise-38.php -- 06/21/2021 MF
+    if (value - Math.floor(value) !== 0) {
+
+      return false;
+
+    } else {
+
+      return true;
+
+    };
+
+  };
+
+};
+
+
+export const hasDecimalPlaces = (value, decimalPlaces) => {
+
+  // * Make sure that the value isn't empty. isNaN counts empty values as a number. -- 09/11/2023 MF
+  if (isEmpty(value) || isNaN(formatTrim(value))) {
+
+    return false;
+
+  } else {
+
+    let currentDecimalPlaces = 1;
+
+    if (Number.isInteger(parseFloat(decimalPlaces))) {
+
+      currentDecimalPlaces = decimalPlaces;
+
+    };
+
+    // * This removes any values from the string starting at and after a non-number value in the string. -- 06/21/2021 MF
+    // * Parse the value to see if it is a float. -- 06/21/2021 MF
+    let valueToTest = parseFloat(formatTrim(value));
+
+    let valueDecimals = 0;
+
+    if (formatTrim(value).includes(".")) {
+
+      // * Remove the characters after the decimal point to be counted later if there is a decimal point. -- 06/21/2021 MF
+      valueDecimals = formatTrim(value).substring(formatTrim(value).indexOf(".") + 1);
+
+    };
+
+    if (isEmpty(valueToTest) || isNaN(valueToTest) || (!isEmpty(currentDecimalPlaces) && valueDecimals.length > currentDecimalPlaces)) {
+
+      return false;
+
+    } else {
+
+      return true;
+
+    };
+
+  };
+
+};
+
+
+export const generateRandomNumber = (minimumValue, maximumValue) => {
+
+  // * https://stackoverflow.com/questions/4959975/generate-random-number-between-two-numbers-in-javascript -- 01/14/2022 MF
+
+  let randomNumber = Math.floor(Math.random() * (maximumValue - minimumValue + 1)) + minimumValue;
+
+  return randomNumber;
+
+};
+
+
+export const generateRandomNumberDigits = (digits) => {
+
+  let randomNumber = formatToString(Math.floor(Math.random() * 10 ** digits));
+
+  while (randomNumber.length < (digits)) {
+
+    randomNumber = `0${randomNumber}`;
+
+  };
+
+  return randomNumber;
+
+};
+
+
+export const formatPhoneNumber = (phoneNumber) => {
+
+  // * From https://learnersbucket.com/examples/javascript/how-to-format-phone-number-in-javascript/ -- 07/22/2021 MF
+
+  let onlyDigits = "";
+
+  if (typeof phoneNumber === "string") {
+
+    onlyDigits = phoneNumber.replace(/\D/g, "");
+
+  };
+
+  let validPhoneNumber = onlyDigits.match(/^(\d{3})(\d{3})(\d{4})$/);
+
+  if (isNonEmptyArray(validPhoneNumber)) {
+
+    return `${validPhoneNumber[1]}-${validPhoneNumber[2]}-${validPhoneNumber[3]}`;
+
+  } else {
+
+    return phoneNumber;
+
+  };
+
+};
+
+
+export const formatTitle = (title) => {
+
+  // * From https://stackoverflow.com/questions/11427759/how-to-insert-space-before-capitalize-character-in-a-word-using-replace-and-rege -- 08/10/2021 MF
+  // * From https://attacomsian.com/blog/string-capitalize-javascript -- 08/10/2021 MF
+  // * From https://www.codeproject.com/Articles/108996/Splitting-Pascal-Camel-Case-with-RegEx-Enhancement -- 08/10/2021 MF
+
+  let formattedTitle = "";
+
+  // * iSBARs is the special case that is difficult to make work in regex. -- 08/16/2021 MF
+  if (!isEmpty(title) && title !== "iSBAR" && title !== "iSBARs" && title !== "iSBAREnable") {
+
+    formattedTitle = title.replace(/(?<!^)([A-Z][a-z]|(?<=[a-z])[A-Z])/g, "$1").replace(/\b\w/g, c => c.toUpperCase());
+
+    // * Accounts for 's in strings. -- 10/03/2024 MF
+    formattedTitle = formattedTitle.replaceAll("'S", "'s");
+
+  } else if (!isEmpty(title) && title === "iSBAR") {
+
+    formattedTitle = "iSBAR";
+
+  } else if (!isEmpty(title) && title === "iSBARs") {
+
+    formattedTitle = "iSBARs";
+
+  } else if (!isEmpty(title) && title === "iSBAREnable") {
+
+    formattedTitle = "iSBAR Enable";
+
+  };
+
+  return formattedTitle;
+
+};
+
+
+export const randomizeItems = (items, randomize) => {
+
+  let itemsRandomized = [];
+
+  if (randomize === true && isNonEmptyArray(items)) {
+
+    itemsRandomized = items.map((a) => {
+
+      return (
+        { sort: Math.random(), value: a }
+      );
+
+    })
+      .sort((a, b) => a.sort - b.sort)
+      .map((a) => a.value);
+
+  } else {
+
+    itemsRandomized = items;
+
+  };
+
+  return itemsRandomized;
+
+};
+
+
+export const getObjectArrayUniqueProperty = (objectArray, uniqueProperty) => {
+
+  // * https://stackoverflow.com/questions/15125920/how-to-get-distinct-values-from-an-array-of-objects-in-javascript -- 05/07/2022 MF
+
+  let uniqueArray = [...objectArray];
+
+  if (isNonEmptyArray(uniqueArray)) {
+
+    uniqueArray = [...new Set(objectArray.map((item) => item[uniqueProperty]))];
+
+    if (typeof uniqueArray[0] === "number") {
+
+      uniqueArray.sort(function (a, b) { return a - b; });
+
+    } else {
+
+      uniqueArray.sort();
+
+    };
+
+  };
+
+  return uniqueArray;
+
+};
+
+
+export const removeArticlesFromBeginning = (value) => {
+
+  let newValue = value;
+
+  if (!isEmpty(value)) {
+
+    newValue = formatLowerCase(newValue).replace(/^(a\.)/, "").replace(/^(an\.)/, "").replace(/^(the\.)/, "");
+
+  };
+
+  return newValue;
+
+};
+
+
+export const compareItemsForSorting = (itemOne, itemTwo) => {
+
+  if (typeof itemOne === "number") {
+
+    return itemOne - itemTwo;
+
+  } else {
+
+    return removeArticlesFromBeginning(itemOne) > removeArticlesFromBeginning(itemTwo);
+
+  };
+
+};
+
+
+export const sortObjectArrayByProperty = (objectArray, sortProperty, direction) => {
+
+  let sortedArray = [...objectArray];
+
+  if (isNonEmptyArray(sortedArray)) {
+
+    if (!isEmpty(sortProperty)) {
+
+      // * https://typeofnan.dev/sort-array-objects-multiple-properties-javascript/ -- 01/07/2025 JH
+
+      sortedArray.sort((a, b) => {
+
+        let aProperty = typeof a[sortProperty] === "number" ? a[sortProperty] : removeArticlesFromBeginning(a[sortProperty]);
+        let bProperty = typeof b[sortProperty] === "number" ? b[sortProperty] : removeArticlesFromBeginning(b[sortProperty]);
+
+        // * Put null values at the end of the array: https://stackoverflow.com/a/29829361 -- 11/30/2023 JH
+        // * The return 0 at the end should cover this case. -- 01/09/2025 JH
+        // if (aProperty === bProperty) return 0;
+
+        if (isEmpty(aProperty)) return 1;
+
+        if (isEmpty(bProperty)) return -1;
+
+        if (aProperty < bProperty) return -1;
+
+        if (aProperty > bProperty) return 1;
+
+        return 0;
+
+      });
+
+    };
+
+  };
+
+  if (formatLowerCase(direction) === "desc") {
+
+    sortedArray.reverse();
+
+  };
+
+  return sortedArray;
+
+};
+
+
+export const sortObjectArrayByTwoProperties = (objectArray, sortPropertyOne, sortPropertyTwo, directionOne, directionTwo) => {
+
+  let sortedArray = [...objectArray];
+
+  if (isNonEmptyArray(sortedArray)) {
+
+    if (!isEmpty(sortPropertyTwo)) {
+
+      sortedArray = sortObjectArrayByProperty(sortedArray, sortPropertyTwo, directionTwo);
+
+    };
+
+    if (!isEmpty(sortPropertyOne)) {
+
+      sortedArray = sortObjectArrayByProperty(sortedArray, sortPropertyOne, directionOne);
+
+    };
+
+  };
+
+  return sortedArray;
+
+};
+
+
+export const compareObjectProperties = (originalObject, comparisonObject) => {
+
+  // * https://stackoverflow.com/questions/14368596/how-can-i-check-that-two-objects-have-the-same-set-of-property-names -- 05/18/2022 MF
+  // * The order of the objects in the parameters matters because the comparison is completed based on what does or does not exist in the comparisionObject. -- 05/18/2022 MF
+
+  let originalObjectProperties = Object.keys(originalObject);
+  let comparisonObjectProperties = Object.keys(comparisonObject);
+
+  let newProperties = [];
+  let removedProperties = [];
+  let sameProperties = [];
+
+  // * Checks for object properties have been added. -- 05/16/2022 MF
+  originalObjectProperties.map((property) => {
+
+    if (comparisonObjectProperties.indexOf(property) < 0) {
+
+      newProperties.push(property);
+
+    } else {
+
+      sameProperties.push(property);
+
+    };
+
+  });
+
+  // // * Checks for object properties have been added. -- 05/16/2022 MF
+  // for (let i = 0; i < originalObjectProperties.length; i++) {
+
+  //   if (comparisonObjectProperties.indexOf(originalObjectProperties[i]) < 0) {
+
+  //     newProperties.push(originalObjectProperties[i]);
+
+  //   } else {
+
+  //     sameProperties.push(originalObjectProperties[i]);
+
+  //   };
+
+  // };
+
+  // * Checks for object properties have been removed. -- 05/16/2022 MF
+  comparisonObjectProperties.map((property) => {
+
+    if (originalObjectProperties.indexOf(property) < 0) {
+
+      removedProperties.push(property);
+
+      // } else {
+
+      //   sameProperties.push(property);
+
+    };
+
+  });
+
+  // // * Checks for object properties have been removed. -- 05/16/2022 MF
+  // for (let i = 0; i < comparisonObjectProperties.length; i++) {
+
+  //   if (originalObjectProperties.indexOf(comparisonObjectProperties[i]) < 0) {
+
+  //     newProperties.push(comparisonObjectProperties[i]);
+
+  //     // } else {
+
+  //     //   sameProperties.push(comparisonObjectProperties[i]);
+
+  //   };
+
+  return { newProperties: [...newProperties], removedProperties: [...removedProperties], sameProperties: [...sameProperties] };
+
+};
+
+
+export const groupObjectArrayByProperties = (objectArray, ...keys) => {
+
+  // * From https://gist.github.com/robmathers/1830ce09695f759bf2c4df15c29dd22d?permalink_comment_id=3646957#gistcomment-3646957 -- 04/04/2022 KH
+
+  const getGroupFromItem = (item, keys) => {
+
+    return (keys.length > 1) ? getGroupFromItem(item[keys[0]], keys.slice(1)) : item[keys[0]];
+
+  };
+
+  return objectArray.reduce((results, item) => {
+
+    // * Get the first instance of the key by which we're grouping. -- 04/04/2022
+    let group = getGroupFromItem(item, keys);
+
+    // * Ensure that there's an array to hold our results for this group. -- 04/04/2022
+    results[group] = results[group] || [];
+
+    // * Add this item to the appropriate group within results. -- 04/04/2022
+    results[group].push(item);
+
+    // * Return the updated results object to be passed into next reduce call. -- 04/04/2022
+    return results;
+
+    // * Initial value of the results object -- 04/04/2022
+  }, {});
+
+};
+
+
+export const formatLowerCase = (value) => {
+
+  let lowerCaseValue = "";
+
+  if (!isEmpty(value)) {
+
+    lowerCaseValue = formatToString(value).toLowerCase();
+
+  };
+
+  return lowerCaseValue;
+
+};
+
+
+export const formatUpperCase = (value) => {
+
+  let upperCaseValue = "";
+
+  if (!isEmpty(value)) {
+
+    upperCaseValue = formatToString(value).toUpperCase();
+
+  };
+
+  return upperCaseValue;
+
+};
+
+
+export const formatTrim = (value) => {
+
+  let trimValue = "";
+
+  if (!isEmpty(value)) {
+
+    trimValue = formatToString(value).trim();
+
+  };
+
+  return trimValue;
+
+};
+
+
+export const formatToString = (value) => {
+
+  let toStringValue = "";
+
+  if (!isEmpty(value)) {
+
+    toStringValue = value.toString();
+
+  };
+
+  return toStringValue;
+
+};
+
+
+export const getParseInt = value => {
+
+  return !isEmpty(value) && !Number.isNaN(value) ? parseInt(value) : null;
+
+};
+
+
+export const formatInt = (value) => {
+
+  let formatedInt = "";
+
+  if (!isEmpty(value)) {
+
+    formatedInt = getParseInt(formatTrim(value.replaceAll(",", ""))).toLocaleString();
+
+  };
+
+  return formatedInt;
+
+};
+
+
+export const formatFloat = (value) => {
+
+  let formatedFloat = "";
+
+  if (!isEmpty(value)) {
+
+    formatedFloat = parseFloat(formatTrim(value.replaceAll(",", ""))).toLocaleString();
+
+  };
+
+  return formatedFloat;
+
+};
+
+
+export const formatDate = (dateToFormat) => {
+
+  let formattedDate = "";
+
+  if (!isEmpty(dateToFormat)) {
+
+    formattedDate = dateToFormat.substring(0, 10);
+
+  } else {
+
+    formattedDate = "";
+
+  };
+
+  return formattedDate;
+
+};
+
+
+export const formatSearchInput = (value) => {
+
+  let formatedSearchInput = "";
+
+  if (!isEmpty(value)) {
+
+    formatedSearchInput = formatTrim(value).toLowerCase();
+
+  };
+
+  return formatedSearchInput;
+
+};
+
+
+export const removeHTML = (text) => {
+
+  // * https://www.geeksforgeeks.org/how-to-strip-out-html-tags-from-a-string-using-javascript/ -- 02/08/2022 MF
+
+  let displayText = "";
+
+  if (!isEmpty(text)) {
+
+    displayText = text.replace(/(<([^>]+)>)/ig, "");
+
+  };
+
+  return displayText;
+
+};
+
+
+export const removeNonAlphanumericCharacters = (text) => {
+
+  // * Removes all characters that aren't letters, numbers, spaces or a period. -- 05/12/2022 MF
+
+  let formatedText = "";
+
+  if (!isEmpty(text)) {
+
+    formatedText = text.replace(/[^a-zA-Z0-9\. ]/g, "");
+
+  };
+
+  return formatedText;
+
+};
+
+
+export const replaceSmartCharacters = (jsonData) => {
+
+  let newJSON = jsonData;
+
+  if (!isEmpty(newJSON)) {
+
+    newJSON = newJSON.replaceAll("’", "'");
+
+    // newJSON = newJSON.replaceAll("–", "--");
+    newJSON = newJSON.replaceAll("–", "-");
+
+    newJSON = newJSON.replaceAll(" ", " ");
+
+    newJSON = newJSON.replaceAll("“", "\"");
+    newJSON = newJSON.replaceAll("”", "\"");
+
+  };
+
+  return newJSON;
+
+};
+
+
+export const getQueryStringData = () => {
+
+  let queryStringData = {};
+
+  // * Retreive the queryString values if there are any. -- 01/22/2023 MF
+  if (typeof window !== "undefined") {
+
+    let queryStrings = new URLSearchParams(window.location.search);
+
+    queryStringData.parametersURL = queryStrings.toString();
+
+    // * From https://medium.com/swlh/urlsearchparams-in-javascript-df524f705317 -- 01/22/2023 MF
+    queryStrings.forEach(function (value, key) {
+
+      // console.log("key", key);
+      // console.log("value", value);
+
+      queryStringData[key] = value;
+
+    });
+
+  };
+
+  return queryStringData;
+
+};
+
+
+export const addLog = (baseURL, fetchAuthorization, databaseAvailable, allowLogging, logObject) => {
+
+  let logResult = "Add log not attempted due to parameter values.";
+
+  if (allowLogging === true && databaseAvailable !== false) {
+
+    let operation = "Add Log";
+
+    let url = `${baseURL}logs/`;
+    let response = "";
+    let data = "";
+
+    fetch(url, {
+      method: "POST",
+      headers: new Headers({
+        "Content-Type": "application/json", "Authorization": fetchAuthorization
+      }),
+      body: JSON.stringify({ recordObject: logObject })
+    })
+      .then(results => {
+
+        response = results;
+
+        if (response.status === 200) {
+
+          return response.json();
+
+        } else {
+
+          addErrorLog(baseURL, fetchAuthorization, databaseAvailable, allowLogging, { operation: `${operation} SQL Server`, transactionData: { url, response: { ok: response.ok, redirected: response.redirected, status: response.status, statusText: response.statusText, type: response.type, url: response.url }, data, logObject: logObject }, errorData: { message: `${response.status} ${response.statusText} ${response.url}` }, dateEntered: getDateTime() });
+
+          logResult = `${operation}: ${response.status} ${response.statusText} ${response.url}`;
+
+        };
+
+      })
+      .then(results => {
+
+        data = results;
+
+        logResult = results;
+
+      })
+      .catch((error) => {
+
+        // console.error("addLog error", error);
+        // console.error("addLog error.name", error.name);
+        // console.error("addLog error.message", error.message);
+        // console.error("addLog error.stack", error.stack);
+
+        // dispatch(addErrorMessage(`${operation}: ${convertSpecialCharacters(error.name)}: ${convertSpecialCharacters(error.message)}`));
+
+        addErrorLog(baseURL, fetchAuthorization, databaseAvailable, allowLogging, { operation, transactionData: { url, response: { ok: response.ok, redirected: response.redirected, status: response.status, statusText: response.statusText, type: response.type, url: response.url }, data, logObject }, errorData: { name: error.name, message: error.message, stack: error.stack }, dateEntered: getDateTime() });
+
+        logResult = `${operation}: ${convertSpecialCharacters(error.name)}: ${convertSpecialCharacters(error.message)}`;
+
+      });
+
+  };
+
+  return logResult;
+
+};
+
+
+export const addErrorLog = (baseURL, fetchAuthorization, databaseAvailable, allowLogging, errorObject) => {
+
+  let logErrorResult = "Add error log not attempted due to parameter values.";
+
+  if (allowLogging === true && databaseAvailable !== false) {
+
+    let operation = "Add Error Log";
+
+    let url = `${baseURL}errorLogs/`;
+    let response = "";
+    // let data = "";
+
+    fetch(url, {
+      method: "POST",
+      headers: new Headers({
+        "Content-Type": "application/json", "Authorization": fetchAuthorization
+      }),
+      body: JSON.stringify({ recordObject: errorObject })
+    })
+      .then(results => {
+
+        response = results;
+
+        if (response.status === 200) {
+
+          return response.json();
+
+        } else {
+
+          logResult = `${operation}: ${response.status} ${response.statusText} ${response.url}`;
+
+        };
+
+      })
+      .then(results => {
+
+        // data = results;
+
+        logErrorResult = results;
+
+      })
+      .catch((error) => {
+
+        // console.error("addErrorLog error", error);
+        // console.error("addErrorLog error.name", error.name);
+        // console.error("addErrorLog error.message", error.message);
+
+        // addErrorMessage(`${operation}: ${convertSpecialCharacters(error.name)}: ${convertSpecialCharacters(error.message)}`);
+
+        logErrorResult = `${operation}: ${convertSpecialCharacters(error.name)}: ${convertSpecialCharacters(error.message)}`;
+
+      });
+
+  };
+
+  return logErrorResult;
+
+};
+
+
+export const addComputerLog = (computerLogOne, computerLogTwo) => {
+
+  let computerLog = { ...computerLogOne };
+
+  if (typeof computerLogItem === "object") {
+
+    // * From https://geolocation-db.com/json/ -- 09/27/2021 MF
+    if (!isEmpty(computerLogTwo.country_code)) {
+
+      computerLog.countryCode = computerLogTwo.country_code;
+
+    };
+
+    if (!isEmpty(computerLogTwo.country_name)) {
+
+      computerLog.countryName = computerLogTwo.country_name;
+
+    };
+
+    if (!isEmpty(computerLogTwo.city)) {
+
+      computerLog.city = computerLogTwo.city;
+
+    };
+
+    if (!isEmpty(computerLogTwo.postal)) {
+
+      computerLog.postal = computerLogTwo.postal;
+
+    };
+
+    if (!isEmpty(computerLogTwo.latitude)) {
+
+      computerLog.latitude = computerLogTwo.latitude;
+
+    };
+
+    if (!isEmpty(computerLogTwo.longitude)) {
+
+      computerLog.longitude = computerLogTwo.longitude;
+
+    };
+
+    if (!isEmpty(computerLogTwo.IPv4)) {
+
+      computerLog.ipAddress = computerLogTwo.IPv4;
+
+    };
+
+    if (!isEmpty(computerLogTwo.state)) {
+
+      computerLog.state = computerLogTwo.state;
+
+    };
+
+    // * From https://api.db-ip.com/v2/free/self -- 09/27/2021 MF
+    if (!isEmpty(computerLogTwo.ipAddress)) {
+
+      computerLog.ipAddress = computerLogTwo.ipAddress;
+
+    };
+
+    if (!isEmpty(computerLogTwo.continentCode)) {
+
+      computerLog.continentCode = computerLogTwo.continentCode;
+
+    };
+
+    if (!isEmpty(computerLogTwo.continentName)) {
+
+      computerLog.continentName = computerLogTwo.continentName;
+
+    };
+
+    if (!isEmpty(computerLogTwo.countryCode)) {
+
+      computerLog.countryCode = computerLogTwo.countryCode;
+
+    };
+
+    if (!isEmpty(computerLogTwo.countryName)) {
+
+      computerLog.countryName = computerLogTwo.countryName;
+
+    };
+
+    if (!isEmpty(computerLogTwo.stateProvCode)) {
+
+      computerLog.stateProvCode = computerLogTwo.stateProvCode;
+
+    };
+
+    if (!isEmpty(computerLogTwo.stateProv)) {
+
+      computerLog.state = computerLogTwo.state;
+
+    };
+
+    if (!isEmpty(computerLogTwo.city)) {
+
+      computerLog.city = computerLogTwo.city;
+
+    };
+
+  };
+
+  return computerLog;
+
+};
+
+
+export const parse = (value, options) => {
+
+  // * The parseHTML function from the npm package html-react-parser doesn't provide error handling if the value sent to it isn't a string. -- 03/09/2023 MF
+
+  let newValue = value;
+
+  if (!isEmpty(value)) {
+
+    if (!isEmpty(options)) {
+
+      newValue = parseHTML(value, options);
+
+    } else {
+
+      newValue = parseHTML(value);
+
+    };
+
+  };
+
+  return newValue;
+
+};
+
+
+export const displayTime = (dateToDisplay, removeLeadingZeroes) => {
+
+  let newDisplayTime = "";
+
+  if (!isEmpty(dateToDisplay)) {
+
+    // * Time
+    let time = formatToString(dateToDisplay).substring(11, 16);
+
+    newDisplayTime = time;
+
+    if (!isEmpty(newDisplayTime) && removeLeadingZeroes === true) {
+
+      newDisplayTime = newDisplayTime.replace(/\b0/g, "");
+
+    };
+
+  };
+
+  return newDisplayTime;
+
+};
+
+
+export const convertMilitaryTimeToStandardTime = (timeEntered) => {
+
+  // * https://stackoverflow.com/questions/29206453/best-way-to-convert-military-time-to-standard-time-in-javascript -- 09/18/2023 KH
+
+  // * timeEntered must be a string in HH:MM format. -- 09/18/2023 KH
+  // * The timeEntered format is no longer a requirement. The function returns an empty string if the format isn't correct. -- 04/23/2025 MF
+
+  let hours = "";
+  let minutes = "";
+  let modifier = "";
+
+  let standardTime = "";
+
+  // * Split the time by : -- 11/27/2023 KH
+  let newTime = !isEmpty(timeEntered) ? timeEntered.split(/[\s:]+/) : [];
+
+  hours = !isEmpty(newTime[0]) ? Number(newTime[0]) : "";
+
+  minutes = !isEmpty(newTime[1]) ? Number(newTime[1]) : "";
+
+  if (hours > 0 && hours <= 12) {
+
+    standardTime = "" + hours;
+
+  } else if (hours > 12) {
+
+    standardTime = "" + (hours - 12);
+
+  } else if (hours === 0) {
+
+    standardTime = "12";
+
+  };
+
+  if (hours >= 12) {
+
+    modifier = " PM";
+
+  } else if (hours > 0 && hours < 12) {
+
+    modifier = " AM";
+
+  };
+
+  if (!isEmpty(standardTime)) {
+
+    standardTime += (minutes > -1 && minutes < 10) ? ":0" + minutes + modifier : ":" + minutes + modifier;
+
+  };
+
+  return standardTime;
+
+};
+
+
+export const convertStandardTimeToMilitaryTime = (timeEntered) => {
+
+  // * https://www.tutorialspoint.com/converting-12-hour-format-time-to-24-hour-format-in-javascript -- 11/27/2023 KH
+
+  // * timeEntered must be a string in HH:MM AM/PM format. -- 09/18/2023 KH
+  // * The timeEntered format is no longer a requirement. The function returns an empty string if the format isn't correct. -- 04/23/2025 MF
+
+  let hours = "";
+  let minutes = "";
+  let modifier = "";
+
+  let militaryTime = "";
+
+  // * Split the time by : and " " -- 11/27/2023 KH
+  let newTime = !isEmpty(timeEntered) ? timeEntered.split(/[\s: ]+/) : [];
+
+  if (!isEmpty(newTime[0])) {
+
+    hours = Number(newTime[0]);
+    hours = (hours < 10) ? "0" + hours : hours;
+
+  };
+
+  if (!isEmpty(newTime[1])) {
+
+    minutes = Number(newTime[1]);
+    minutes = (minutes < 10) ? "0" + minutes : minutes;
+
+  };
+
+  modifier = !isEmpty(newTime[2]) ? newTime[2] : "";
+
+  if (hours === 12) {
+
+    hours = "00";
+
+  };
+
+  if (modifier === "PM") {
+
+    // * 10 indicates that the string is in base 10 (decimal notation). -- 11/27/2023 KH
+    // * + 12 converts the 12-hour format to a 24-hour format. -- 11/27/2023 KH
+    hours = getParseInt(hours, 10) + 12;
+
+  };
+
+  if (!isEmpty(hours) && !isEmpty(minutes)) {
+
+    militaryTime = `${hours}:${minutes}`;
+
+  };
+
+  return militaryTime;
+
+};
+
+
+export const getNumberOfDaysBetweenDates = (startDate, endDate) => {
+
+  // * https://www.geeksforgeeks.org/how-to-calculate-the-number-of-days-between-two-dates-in-javascript/ -- 09/12/2023 KH
+  // * https://stackoverflow.com/questions/2627473/how-to-calculate-the-number-of-days-between-two-dates/2627493#2627493 -- 09/12/2023 KH
+
+  let newStartDate = new Date(startDate);
+  let newEndDate = new Date(endDate);
+  let numberOfDays = 1;
+
+  // * The function getTime() converts days to milliseconds to subtract the two dates. -- 08/31/2023 KH  
+  // * The equation (1000 * 60 * 60 * 24) turns the values back into a day count. -- 08/31/2023 KH  
+  // * The operation + 1 adds one day to account for 08/31/2023 - 08/31/2023 = 0 but should equal 1 day in order to account for running studentsPerDay amount of students that day. -- 08/31/2023 KH  
+  numberOfDays = ((newEndDate.getTime() - newStartDate.getTime()) / (1000 * 60 * 60 * 24)) + 1;
+
+  return numberOfDays;
+
+};
+
+
+export const convertTimeToMinutes = (timeEntered) => {
+
+  // * https://stackoverflow.com/questions/32885682/convert-hhmmss-into-minute-using-javascript -- 11/27/2023 KH
+  // * timeEntered must be a string in HH:MM format. -- 09/18/2023 KH
+
+  let newTime = timeEntered.split(':');
+
+  let minutes = (+newTime[0]) * 60 + (+newTime[1]);
+
+  return minutes;
+
+};
+
+
+export const generateHoursInterval = (startHourInMinutes, endHourInMinutes, interval) => {
+
+  // * https://gist.github.com/indexzero/6261ad9292c78cf3c5aa69265e2422bf?permalink_comment_id=4003346#gistcomment-4003346 -- 11/20/2023 KH
+
+  let timesArray = [];
+
+  if (!isEmpty(startHourInMinutes) && isWholeNumber(startHourInMinutes)) {
+
+    for (let i = 0; startHourInMinutes < 24 * 60; i++) {
+
+      if (startHourInMinutes > endHourInMinutes) {
+
+        break;
+
+      } else {
+
+        // * Get hours of day in 0-24 format. -- 11/20/2023 KH
+        let hh = Math.floor(startHourInMinutes / 60);
+
+        // * Get minutes of the hour in 0-55 format. -- 11/20/2023 KH
+        let mm = startHourInMinutes % 60;
+
+        timesArray[i] = { timeID: i, time: convertMilitaryTimeToStandardTime(("0" + (hh % 24)).slice(-2) + ":" + ("0" + mm).slice(-2)) };
+
+        startHourInMinutes = startHourInMinutes + interval;
+
+      };
+
+    };
+
+  };
+
+  return timesArray;
+
+};
+
+
+export const displayCommaBetweenTwoItems = (firstItem, secondItem) => {
+
+  let textValue = "";
+
+  if (!isEmpty(firstItem)) {
+
+    textValue = firstItem;
+
+  };
+
+  if (!isEmpty(firstItem) && !isEmpty(secondItem)) {
+
+    textValue += ", ";
+
+  };
+
+  if (!isEmpty(secondItem)) {
+
+    textValue += secondItem;
+
+  };
+
+  return textValue;
+
+};
+
+
+export const constructCSVFile = (dataList, dataColumnTitles, propertyNameList) => {
+
+  let fileData = "";
+
+  // * Initialize fileData with columns titles. -- 10/15/2024 EBG
+  if (!isEmpty(dataColumnTitles)) {
+
+    fileData += dataColumnTitles;
+
+  };
+
+  if (isNonEmptyArray(dataList) && isNonEmptyArray(propertyNameList)) {
+
+    // * Loop through every record in the data. -- 10/16/2024 EBG
+    for (let i = 0; i < dataList.length; i++) {
+
+      // * Look at each object property name specified in propertyNameList. -- 10/16/2024 EBG
+      for (let j = 0; j < propertyNameList.length; j++) {
+
+        let cellData = "";
+
+        let propertyNameData = dataList[i][propertyNameList[j].propertyName];
+
+        if (!isEmpty(propertyNameData)) {
+
+          // ? Will json need its own case or be passed as a string? -- 10/17/2024 EBG
+          // * When JSON is passed, I'm pretty sure it's going to be a different case in these if statements. -- 10/17/2024 MF
+          if (propertyNameList[j].dataType === "string") {
+
+            cellData = `"${propertyNameData}"`;
+
+          } else if (propertyNameList[j].dataType === "number") {
+
+            cellData = `${propertyNameData}`;
+
+          } else if (propertyNameList[j].dataType === "boolean") {
+
+            cellData = `${propertyNameData === true || propertyNameData === 1 ? "Yes" : "-"}`;
+
+          } else if (propertyNameList[j].dataType === "date") {
+
+            cellData = `${displayDate(propertyNameData)}`;
+
+          } else if (propertyNameList[j].dataType === "dateTime") {
+
+            cellData = `${displayDateAndTime(propertyNameData)}`;
+
+          } else if (propertyNameList[j].dataType === "hyperlink") {
+
+            // * Creates a hyperlink in the CSV file. -- 05/30/2024 MF
+            // * https://stackoverflow.com/questions/7572268/how-to-encode-a-hyperlink-in-csv-formatted-file -- 05/30/2024 MF
+
+            cellData = "=HYPERLINK(" + `${propertyNameData}` + ")";
+
+          };
+
+        };
+
+        fileData += `${cellData},`;
+
+      };
+
+      fileData += `\r\n`;
+
+    };
+
+  };
+
+  return fileData;
+
+};
+
+
+export const exportCSVFile = (fileData, fileTitle) => {
+
+  // * https://medium.com/@gb.usmanumar/how-to-export-data-to-csv-json-in-react-js-ea45d940652a -- 05/30/2024 EBG
+
+  const csvData = new Blob([fileData], { type: "text/csv" });
+  const csvURL = window.URL.createObjectURL(csvData);
+  const link = document.createElement("a");
+
+  link.href = csvURL;
+  link.download = `${fileTitle} ${displayDate(getDateTime()).replaceAll("/", "_")}.csv`;
+
+  document.body.appendChild(link);
+  link.click();
+  document.body.removeChild(link);
+
+};
+
+
+// export const convertValueToBoolean = (value) => {
+
+//   let stringValue = formatLowerCase(value);
+
+//   switch (stringValue) {
+//     case "true":
+//       // case "yes":
+//       // case "enable":
+//       // case "normal":
+//       return true;
+//     case "True":
+//       return true;
+//     case "1":
+//       return true;
+//     case 1:
+//       return true;
+//     case "false":
+//       // case "no":
+//       // case "disable":
+//       // case "abnormal":
+//       return false;
+//     case "False":
+//       return false;
+//     case "0":
+//       // case null:
+//       // case undefined:
+//       return false;
+//     case 0:
+//       return false;
+//     default:
+//       return value;
+//   };
+
+// };
+
+
+export const convertMinutesToSeconds = (time) => {
+
+  let timeInSeconds = 0;
+
+  if (!isEmpty(time)) {
+
+    let splitTime = time.split(/[ :]/);
+
+    let minute = getParseInt(splitTime[0]) * 60;
+    let second = getParseInt(splitTime[1]);
+
+    timeInSeconds = minute + second;
+
+  };
+
+  return timeInSeconds;
+
+};
+
+
+// export const displaySecondsAsMinutes = (seconds) => {
+
+//   let newTime = "";
+
+//   if (!isEmpty(seconds)) {
+
+//     let newMinutes = Math.floor((getParseInt(seconds) % 3600) / 60);
+//     let newSeconds = getParseInt(seconds) % 60;
+
+//     // newMinutes = newMinutes.toString().padStart(2, "0");
+//     newSeconds = newSeconds.toString().padStart(2, "0");
+
+//     newTime = `${newMinutes}:${newSeconds}`;
+
+//   };
+
+//   return newTime;
+
+// };
+
+
+export const createDateFromString = (dateString) => {
+
+  let [year, month, day] = dateString.split("-").map(Number);
+
+  if (dateString.includes("/")) {
+
+    [month, day, year] = dateString.split("/").map(Number);
+
+  };
+
+  return new Date(year, month - 1, day);  // * Month is 0-indexed. -- 04/08/2025 MF
+
+};
+
+
+export const isFutureDate = (dateToCheck, referenceDate) => {
+
+  const dateOne = new Date(dateToCheck.getFullYear(), dateToCheck.getMonth(), dateToCheck.getDate());
+
+  const dateTwo = new Date(referenceDate.getFullYear(), referenceDate.getMonth(), referenceDate.getDate());
+
+  return dateOne > dateTwo;
+
+};
+
+
+export const formatToUSD = (value) => {
+
+  if (!Number.isNaN(value)) {
+
+    let formattedValue = new Intl.NumberFormat("en-US", {
+      style: "currency",
+      currency: "USD",
+      maximumFractionDigits: 0
+    }).format(value);
+
+    return formattedValue;
+
+  } else {
+
+    return "";
+
+  };
+
+};
+
+
+export const getYears = (months) => {
+
+  let newValue = Number.parseFloat(months / 12).toFixed(1);
+
+  if (!isNaN(newValue)) {
+
+    return newValue;
+
+  } else {
+
+    return "";
+
+  };
+
+};
+
+
+export const parseDescription = (description) => {
+
+  // console.log("description", description.split("\\r"));
+
+  return parse(description.replace(/\\r/g, "<br />"));
+  // return parse(description.split("\r").join("<br />"));
+
+};
+
+
+export const getBrowserData = () => ({ appCodeName: navigator.appCodeName, appName: navigator.appName, appVersion: navigator.appVersion, cookieEnabled: navigator.cookieEnabled, language: navigator.language, onLine: navigator.onLine, platform: navigator.platform, product: navigator.product, userAgent: navigator.userAgent });
+
+
+export const returnActiveClass = (componentName, componentToLoad, classList) => {
+
+  let newClassList = !isEmpty(classList) ? classList : "";
+
+  newClassList += componentName === componentToLoad ? " active" : "";
+
+  return newClassList;
+
+};
+
+
+// ? removeArrayItemsByProperty -- 11/14/2025 MF
+export const filterDropdownOptions = (allItems, itemsToFilter, propertyName, currentItem) => {
+
+  let filteredItems = [];
+
+  if (!isEmptyArray(allItems)) {
+
+    // * https://stackoverflow.com/a/71904924 -- 07/28/2025 JH
+    filteredItems = allItems.filter(item => !itemsToFilter.find(itemToFilter => item[propertyName] === itemToFilter[propertyName]));
+
+    if (!isEmpty(currentItem)) {
+
+      filteredItems.push(currentItem);
+
+    };
+
+  };
+
+  return filteredItems;
+
+};
+
+
+// ? filterArrayByPropertyArray -- 11/14/2025 MF
+// ? Is this function the same as filterDropdownOptions? -- 11/14/2025 MF
+// TODO: Found two filterBySelectedCheckbox functions. This one is located in Extended Reality Scheduling and Partners. -- 11/14/2025 MF
+export const filterBySelectedCheckbox = (checkboxValues, selectedID, currentResults) => {
+
+  let filteredResults = [];
+
+  if (!isEmptyArray(checkboxValues) && !isEmptyArray(currentResults)) {
+
+    checkboxValues.forEach(checkboxValue => {
+
+      let newFilteredResults = currentResults.filter((result) => formatToString(result[selectedID]) === formatToString(checkboxValue));
+
+      filteredResults = filteredResults.concat(newFilteredResults);
+
+    });
+
+  };
+
+  return filteredResults;
+
+};
+
+
+// ? filterArrayByPropertyArray -- 11/14/2025 MF
+// TODO: Found two filterBySelectedCheckbox functions. This one is located in Faculty Training and Learning Object Repository. -- 11/14/2025 MF
+// export const filterBySelectedCheckbox = (checkboxValues, selectedID, currentResults) => {
+
+//   let filteredResults = [];
+
+//   if (isNonEmptyArray(checkboxValues) && isNonEmptyArray(currentResults)) {
+
+//     // ? Two filters? Filter then map? -- 08/24/2023 JH
+//     for (let i = 0; i < checkboxValues.length; i++) {
+
+//       let newFilteredResults = currentResults.filter((item) => formatToString(item[selectedID]) === checkboxValues[i]);
+
+//       filteredResults = filteredResults.concat(newFilteredResults);
+
+//     };
+
+//   };
+
+//   return filteredResults;
+
+// };
+
+
+export const getFilteredNames = (selectedCheckbox, selectedID, optionData, nameToReturn, filteredNames) => {
+
+  let newFilteredNames = filteredNames;
+
+  if (!isEmptyArray(selectedCheckbox) && !isEmptyArray(optionData)) {
+
+    for (let i = 0; i < selectedCheckbox.length; i++) {
+
+      for (let j = 0; j < optionData.length; j++) {
+
+        if (formatToString(optionData[j][selectedID]) === formatToString(selectedCheckbox[i])) {
+
+          if (!isEmpty(newFilteredNames)) {
+
+            newFilteredNames += `, ${optionData[j][nameToReturn]}`;
+
+          } else {
+
+            newFilteredNames = `${optionData[j][nameToReturn]}`;
+
+          };
+
+        };
+
+      };
+
+    };
+
+  };
+
+  return newFilteredNames;
+
+};
+
+
+export const isValidURL = (value) => {
+
+  try {
+
+    new URL(value);
+    return true;
+
+  } catch {
+
+    return false;
+
+  }
+
+};
